requirements.txt
leveldb
slacktokens
pyproject-toml
python-dotenv
slack_sdk
pandas
reportlab